nilfs-utils: kill warnings on nilfs_open_sem
Fixes the following gcc warnings:

 nilfs.c:280: warning: format '%lu' expects type 'long unsigned int', but argument 4 has type '__dev_t'
 nilfs.c:285: warning: format '%lu' expects type 'long unsigned int', but argument 5 has type '__ino64_t'

+++ lib/nilfs.c
@@ -277,11 +277,13 @@ static int nilfs_open_sem(struct nilfs *nilfs)
 
 	if (S_ISBLK(stbuf.st_mode)) {
 		ret = snprintf(semnambuf, sizeof(semnambuf),
-			       "/nilfs-cleaner-%lu", stbuf.st_rdev);
+			       "/nilfs-cleaner-%llu",
+			       (unsigned long long)stbuf.st_rdev);
 	} else if (S_ISREG(stbuf.st_mode) || S_ISDIR(stbuf.st_mode)) {
 		ret = snprintf(semnambuf, sizeof(semnambuf),
-			       "/nilfs-cleaner-%lu-%lu",
-			       stbuf.st_dev, stbuf.st_ino);
+			       "/nilfs-cleaner-%llu-%llu",
+			       (unsigned long long)stbuf.st_dev,
+			       (unsigned long long)stbuf.st_ino);
 	} else {
 		errno = EINVAL;
 		return -1;
